+++ Sources/Bloc/Bloc/BlocBase.swift
@@ -4,39 +4,6 @@
 
 import Foundation
 
-public protocol Streamable<State>: Sendable {
-
-    associatedtype State: Sendable
-
-    var stream: Stream<State> { get }
-}
-
-public protocol StateStreamable<State>: Streamable {
-
-    var state: State { get }
-}
-
-public protocol StateStreamableSource<State>: StateStreamable, Closable {}
-
-public protocol Closable: Sendable {
-
-    var isClosed: Bool { get }
-
-    func close() async throws
-}
-
-protocol Emittable<_State>: Sendable {
-
-    associatedtype _State: Sendable
-
-    func emit(_ state: _State) throws
-}
-
-public protocol ErrorSink: Closable {
-
-    func addError(_ error: Error)
-}
-
 open class BlocBase<State>: StateStreamableSource, Emittable, ErrorSink, @unchecked Sendable {
 
     // MARK: - Open properties
@@ -94,9 +61,7 @@ open class BlocBase<State>: StateStreamableSource, Emittable, ErrorSink, @unchec
         fatalError("close() has not been implemented")
     }
 
-    // MARK: - Internal methods
-
-    func emit(_ state: State) throws {
+    public func emit(_ state: State) throws {
         fatalError("emit(_:) has not been implemented")
     }
 }
