+++ Sources/Bloc/Bloc/Models/BlocEventSink.swift
@@ -1,0 +1,12 @@
+/*
+ See LICENSE for this package's licensing information.
+*/
+
+import Foundation
+
+public protocol BlocEventSink<Event>: ErrorSink {
+
+    associatedtype Event: Sendable
+
+    func add<E>(_ event: E)
+}
