+++ Sources/Bloc/Bloc/Models/DefaultBlocObserver.swift
@@ -1,0 +1,7 @@
+/*
+ See LICENSE for this package's licensing information.
+*/
+
+import Foundation
+
+struct DefaultBlocObserver: BlocObserver {}
