+++ Sources/Bloc/Bloc/Models/ErrorSink.swift
@@ -1,0 +1,10 @@
+/*
+ See LICENSE for this package's licensing information.
+*/
+
+import Foundation
+
+public protocol ErrorSink: Closable {
+
+    func addError(_ error: Error)
+}
