+++ Sources/Bloc/Bloc/Models/EventMapper.swift
@@ -1,0 +1,7 @@
+/*
+ See LICENSE for this package's licensing information.
+*/
+
+import Foundation
+
+public typealias EventMapper<Event: Sendable> = (Event) throws -> Stream<Event>
