+++ Sources/Bloc/Bloc/Models/StateStreamable.swift
@@ -1,0 +1,10 @@
+/*
+ See LICENSE for this package's licensing information.
+*/
+
+import Foundation
+
+public protocol StateStreamable<State>: Streamable {
+
+    var state: State { get }
+}
