+++ Sources/Bloc/Bloc/Models/StateStreamableSource.swift
@@ -1,0 +1,7 @@
+/*
+ See LICENSE for this package's licensing information.
+*/
+
+import Foundation
+
+public protocol StateStreamableSource<State>: StateStreamable, Closable {}
