+++ Sources/Bloc/Bloc/Models/Streamable.swift
@@ -1,0 +1,12 @@
+/*
+ See LICENSE for this package's licensing information.
+*/
+
+import Foundation
+
+public protocol Streamable<State>: Sendable {
+
+    associatedtype State: Sendable
+
+    var stream: Stream<State> { get }
+}
