+++ Sources/Bloc/Emitter/BlocEmitter.swift
@@ -4,45 +4,7 @@
 
 import Foundation
 
-public protocol Emitter<State>: Sendable {
-
-    associatedtype State: Sendable
-
-    var isDone: Bool { get }
-
-    func onEach<T>(
-        _ stream: Stream<T>,
-        onData: @escaping @Sendable (T) -> Void,
-        onError: (@Sendable (Error) -> Void)?
-    ) async throws
-
-    func forEach<T>(
-        _ stream: Stream<T>,
-        onData: @escaping @Sendable (T) -> State,
-        onError: (@Sendable (Error) -> Void)?
-    ) async throws
-
-    func callAsFunction(_ state: State)
-}
-
-extension Emitter {
-
-    public func forEach<T>(
-        _ stream: Stream<T>,
-        onData: @escaping @Sendable (T) -> Void
-    ) async throws {
-        try await onEach(stream, onData: onData, onError: nil)
-    }
-
-    public func onEach<T>(
-        _ stream: Stream<T>,
-        onData: @escaping @Sendable (T) -> State
-    ) async throws {
-        try await forEach(stream, onData: onData, onError: nil)
-    }
-}
-
-class _Emitter<State>: Emitter, @unchecked Sendable {
+class BlocEmitter<State>: Emitter, @unchecked Sendable {
 
     // MARK: - Internal properties
